Abstract away the driver, providing a lighter-weight DistanceSensor class.
